Add files via upload
The aim of this mini project is to simulate mutual exclusion using SEMAPHORES by
exploiting graphical processing capabilities. This mini project demonstrates the working of
semaphores.The operating system consists of all the resources that a process needs.Whenever
more than one process request for the resources, the OS will allocate resource to the process one
at a time based on first come first serve basis.The other processes requesting for the resources are
kept in the waiting queue.Whenever the first process releases the resource,the other process
waiting in the queue are allocated resources by the OS.With the curiosity about synchronization,
we have made our mind to show how orderly execution of co-operating processes can be
achieved using SEMAPHORE to avoid data inconsistency, with the knowledge of computer
graphics.